# api/swagger.yaml
openapi: "3.0.3"
info:
  title: API Cantinho Dona Maria
  version: "1.0.0"
  description: API para gerenciamento de ADMs, Clientes, Espaços, Reservas e Pagamentos
servers:
  - url: http://localhost:3000

tags:
  - name: ADM
    description: Operações relacionadas aos administradores
  - name: Cliente
    description: Operações relacionadas aos clientes
  - name: Espaço
    description: Operações relacionadas aos espaços
  - name: Reserva
    description: Operações relacionadas às reservas
  - name: Pagamento
    description: Operações relacionadas aos pagamentos

paths:

  /adm:
    get:
      tags:
        - ADM
      summary: Listar todos os ADMs
      responses:
        '200':
          description: Lista de ADMs
        '404':
          description: Nenhum ADM encontrado
        '500':
          description: Erro interno do servidor
    post:
      tags:
        - ADM
      summary: Criar um novo ADM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - cpf
              properties:
                nome:
                  type: string
                  description: Nome completo do ADM
                cpf:
                  type: integer
                  description: CPF do ADM
                datanasc:
                  type: string
                  format: date
                  description: Data de nascimento do ADM
      responses:
        '201':
          description: ADM criado com sucesso
        '400':
          description: Requisição inválida
        '500':
          description: Erro interno do servidor

  /adm/{id}:
    get:
      tags:
        - ADM
      summary: Buscar um ADM pelo ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID do ADM
          schema:
            type: integer
      responses:
        '200':
          description: ADM encontrado
        '404':
          description: ADM não encontrado
        '500':
          description: Erro interno do servidor
    put:
      tags:
        - ADM
      summary: Atualizar ADM
      parameters:
        - name: id
          in: path
          required: true
          description: ID do ADM
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                cpf:
                  type: integer
                datanasc:
                  type: string
                  format: date
      responses:
        '200':
          description: ADM atualizado
        '404':
          description: ADM não encontrado
        '500':
          description: Erro interno do servidor
    delete:
      tags:
        - ADM
      summary: Deletar ADM
      parameters:
        - name: id
          in: path
          required: true
          description: ID do ADM
          schema:
            type: integer
      responses:
        '204':
          description: ADM removido
        '404':
          description: ADM não encontrado
        '500':
          description: Erro interno do servidor

# -------------------------
# Cliente
# -------------------------
  /clientes:
    get:
      tags:
        - Cliente
      summary: Listar todos os clientes
      responses:
        '200':
          description: Lista de clientes
        '404':
          description: Nenhum cliente encontrado
        '500':
          description: Erro interno do servidor
    post:
      tags:
        - Cliente
      summary: Criar cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - cpf
              properties:
                nome:
                  type: string
                  description: Nome completo do cliente
                cpf:
                  type: integer
                  description: CPF do cliente
                datanasc:
                  type: string
                  format: date
                  description: Data de nascimento do cliente
      responses:
        '201':
          description: Cliente criado
        '400':
          description: Requisição inválida
        '500':
          description: Erro interno do servidor

  /clientes/{id}:
    get:
      tags:
        - Cliente
      summary: Buscar cliente pelo ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID do cliente
          schema:
            type: integer
      responses:
        '200':
          description: Cliente encontrado
        '404':
          description: Cliente não encontrado
        '500':
          description: Erro interno do servidor
    put:
      tags:
        - Cliente
      summary: Atualizar cliente
      parameters:
        - name: id
          in: path
          required: true
          description: ID do cliente
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                cpf:
                  type: integer
                datanasc:
                  type: string
                  format: date
      responses:
        '200':
          description: Cliente atualizado
        '404':
          description: Cliente não encontrado
        '500':
          description: Erro interno do servidor
    delete:
      tags:
        - Cliente
      summary: Deletar cliente
      parameters:
        - name: id
          in: path
          required: true
          description: ID do cliente
          schema:
            type: integer
      responses:
        '204':
          description: Cliente removido
        '404':
          description: Cliente não encontrado
        '500':
          description: Erro interno do servidor

# -------------------------
# Espaço
# -------------------------
  /espacos:
    post:
      tags:
        - Espaço
      summary: Criar espaço
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - capacidade
                - preco
              properties:
                nome:
                  type: string
                descricao:
                  type: string
                capacidade:
                  type: integer
                preco:
                  type: number
                tipo:
                  type: string
                diasindisponiveis:
                  type: array
                  items:
                    type: string
                    format: date
      responses:
        '201':
          description: Espaço criado
        '400':
          description: Requisição inválida
        '500':
          description: Erro interno do servidor
    get:
      tags:
        - Espaço
      summary: Listar todos os espaços
      responses:
        '200':
          description: Lista de espaços
        '404':
          description: Nenhum espaço encontrado
        '500':
          description: Erro interno do servidor

  /espacos/{id}:
    get:
      tags:
        - Espaço
      summary: Buscar espaço pelo ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID do espaço
          schema:
            type: integer
      responses:
        '200':
          description: Espaço encontrado
        '404':
          description: Espaço não encontrado
        '500':
          description: Erro interno do servidor
    put:
      tags:
        - Espaço
      summary: Atualizar espaço
      parameters:
        - name: id
          in: path
          required: true
          description: ID do espaço
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                descricao:
                  type: string
                capacidade:
                  type: integer
                preco:
                  type: number
                tipo:
                  type: string
                diasindisponiveis:
                  type: array
                  items:
                    type: string
                    format: date
      responses:
        '200':
          description: Espaço atualizado
        '404':
          description: Espaço não encontrado
        '500':
          description: Erro interno do servidor
    delete:
      tags:
        - Espaço
      summary: Deletar espaço
      parameters:
        - name: id
          in: path
          required: true
          description: ID do espaço
          schema:
            type: integer
      responses:
        '204':
          description: Espaço removido
        '404':
          description: Espaço não encontrado
        '500':
          description: Erro interno do servidor

# -------------------------
# Reserva
# -------------------------
  /reservas:
    get:
      tags:
        - Reserva
      summary: Listar todas as reservas
      responses:
        '200':
          description: Lista de reservas
        '500':
          description: Erro interno do servidor
    post:
      tags:
        - Reserva
      summary: Criar uma reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: string
                  format: date
                id_cliente:
                  type: integer
                id_espaco:
                  type: integer
                id_adm:
                  type: integer
      responses:
        '201':
          description: Reserva criada
        '400':
          description: Requisição inválida
        '500':
          description: Erro interno do servidor

  /reservas/{id}:
    get:
      tags:
        - Reserva
      summary: Obter reserva pelo ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID da reserva
          schema:
            type: integer
      responses:
        '200':
          description: Reserva encontrada
        '404':
          description: Reserva não encontrada
        '500':
          description: Erro interno do servidor
    put:
      tags:
        - Reserva
      summary: Atualizar reserva
      parameters:
        - name: id
          in: path
          required: true
          description: ID da reserva
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: string
                  format: date
                id_cliente:
                  type: integer
                id_espaco:
                  type: integer
                id_adm:
                  type: integer
      responses:
        '200':
          description: Reserva atualizada
        '404':
          description: Reserva não encontrada
        '500':
          description: Erro interno do servidor
    delete:
      tags:
        - Reserva
      summary: Deletar reserva
      parameters:
        - name: id
          in: path
          required: true
          description: ID da reserva
          schema:
            type: integer
      responses:
        '204':
          description: Reserva removida
        '404':
          description: Reserva não encontrada
        '500':
          description: Erro interno do servidor

# -------------------------
# Pagamento
# -------------------------
  /pagamentos:
    get:
      tags:
        - Pagamento
      summary: Listar todos os pagamentos
      responses:
        '200':
          description: Lista de pagamentos
        '500':
          description: Erro interno do servidor
    post:
      tags:
        - Pagamento
      summary: Criar pagamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                valortotal:
                  type: number
                valorpago:
                  type: number
                tipopagamento:
                  type: string
                id_reserva:
                  type: integer
      responses:
        '201':
          description: Pagamento criado
        '400':
          description: Erro ao criar pagamento

  /pagamentos/{id}:
    get:
      tags:
        - Pagamento
      summary: Obter pagamento pelo ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID do pagamento
          schema:
            type: integer
      responses:
        '200':
          description: Pagamento encontrado
        '404':
          description: Pagamento não encontrado
        '500':
          description: Erro interno do servidor
    put:
      tags:
        - Pagamento
      summary: Atualizar pagamento
      parameters:
        - name: id
          in: path
          required: true
          description: ID do pagamento
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                valortotal:
                  type: number
                valorpago:
                  type: number
                tipopagamento:
                  type: string
                id_reserva:
                  type: integer
      responses:
        '200':
          description: Pagamento atualizado
        '404':
          description: Pagamento não encontrado
        '400':
          description: Erro ao atualizar pagamento
    delete:
      tags:
        - Pagamento
      summary: Deletar pagamento
      parameters:
        - name: id
          in: path
          required: true
          description: ID do pagamento
          schema:
            type: integer
      responses:
        '204':
          description: Pagamento removido
        '404':
          description: Pagamento não encontrado
        '500':
          description: Erro interno do servidor